Fix signed/unsigned mismatch compiler warnings. https://bugs.webkit.org/show_bug.cgi?id=80790
Reviewed by Alexey Proskuryakov.

* loader/icon/IconDatabase.cpp:
(WebCore::IconDatabase::synchronousLoadDecisionForIconURL):
* platform/network/MIMESniffing.cpp:


Canonical link: https://commits.webkit.org/98143@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110566 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/loader/icon/IconDatabase.cpp
@@ -652,7 +652,7 @@ IconLoadDecision IconDatabase::synchronousLoadDecisionForIconURL(const String& i
         MutexLocker locker(m_urlAndIconLock);
         if (IconRecord* icon = m_iconURLToRecordMap.get(iconURL)) {
             LOG(IconDatabase, "Found expiration time on a present icon based on existing IconRecord");
-            return (int)currentTime() - icon->getTimestamp() > iconExpirationTime ? IconLoadYes : IconLoadNo;
+            return static_cast<int>(currentTime()) - static_cast<int>(icon->getTimestamp()) > iconExpirationTime ? IconLoadYes : IconLoadNo;
         }
     }
